// src/components/common/head-tag/types.ts
/*
 * FilePath    : blog-client-nuxt\src\components\common\head-tag\types.ts
 * Author      : jiaopengzi
 * Blog        : https://jiaopengzi.com
 * Copyright   : Copyright (c) 2025 by jiaopengzi, All Rights Reserved.
 * Description : 类型
 */

export interface HeadProps {
    title?: string
    description?: string
    keywords?: string
    type?: string
    locale?: string
    author?: string
    image?: string
    siteName?: string
    url?: string
    releaseDate?: string
}